refactor: update articles index view to display a list of articles
This commit refactors the articles index view to display a list of articles. It replaces the previous "Hello, World!" heading with a heading "List of Articles" and adds an unordered list to display each article's title. The @articles instance variable is used to iterate over the articles and display their titles in list items.