feat: add show action to articles controller and show view
This commit adds a new action `show` to the `ArticlesController` and a corresponding view `show.html.erb` in the `articles` directory. The `show` action retrieves an article based on the provided `id` parameter and assigns it to the `@articles` instance variable. The `show.html.erb` view displays the title and body of the article using `<h1>` and `<h2>` tags respectively.