refactor: update articles index view to use link_to helper for article titles
This commit refactors the articles index view to use the `link_to` helper method for generating the anchor tags for article titles. Instead of manually creating the anchor tags with the article URL, the `link_to` method is used to generate the appropriate HTML code. This improves code readability and maintainability.